添加SpringBoot3集成Quartz详细版项目，内容还未填充

+++ 01.SpringBoot/SpringBoot集成组件库/springboot-3-quartz-detailed-demo/src/main/java/com/guoshunfa/springboot3quartzdetaileddemo/controller/JobController.java
@@ -1,71 +1,66 @@
 package com.guoshunfa.springboot3quartzdetaileddemo.controller;
 
 import com.guoshunfa.springboot3quartzdetaileddemo.entity.JobDataMap;
+import com.guoshunfa.springboot3quartzdetaileddemo.service.JobDataMapService;
 import com.guoshunfa.springboot3quartzdetaileddemo.service.QuartzService;
 import com.mongodb.client.result.DeleteResult;
 import com.mongodb.client.result.UpdateResult;
 import jakarta.annotation.Nonnull;
 import jakarta.annotation.Resource;
-import org.springframework.data.mongodb.core.MongoTemplate;
-import org.springframework.data.mongodb.core.query.Criteria;
-import org.springframework.data.mongodb.core.query.Query;
-import org.springframework.data.mongodb.core.query.Update;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 /**
- * job服务调用
+ * 定时任务服务调用
  **/
 @RestController
 public class JobController {
 
     @Resource
-    private MongoTemplate mongoTemplate;
-
-    @Resource
     private QuartzService quartzService;
 
+    @Resource
+    private JobDataMapService jobDataMapService;
+
     /**
-     * 创建/修改job信息
+     * 创建/修改定时任务信息
      *
      * @param jobData
      * @return
      */
     @PostMapping("/job/save")
     public JobDataMap saveJob(@RequestBody JobDataMap jobData) {
-        JobDataMap data = mongoTemplate.save(jobData);
+        JobDataMap data = jobDataMapService.save(jobData);
         quartzService.saveJob(data);
         return data;
     }
 
     /**
-     * 删除job
+     * 删除定时任务
      *
      * @param jobId
      * @return
      */
-    @GetMapping("/job/remove")
+    @DeleteMapping("/job/remove")
     public DeleteResult removeJob(@Nonnull String jobId) {
-        DeleteResult deleteResult = mongoTemplate.remove(Query.query(Criteria.where("id").is(jobId)), JobDataMap.class);
-        JobDataMap data = mongoTemplate.findById(jobId, JobDataMap.class);
-        quartzService.removeJob(data);
+        DeleteResult deleteResult = jobDataMapService.remove(jobId);
+        if (deleteResult.wasAcknowledged()) {
+            JobDataMap data = jobDataMapService.findById(jobId);
+            quartzService.removeJob(data);
+        }
         return deleteResult;
     }
 
     /**
-     * 暂停/运行job
+     * 暂停/运行定时任务
      *
      * @param jobId
      * @param paused
      * @return
      */
-    @GetMapping("/job/paused")
+    @PutMapping("/job/paused")
     public UpdateResult pausedJob(@Nonnull String jobId, @Nonnull Boolean paused) {
-        UpdateResult updateResult = mongoTemplate.updateFirst(Query.query(Criteria.where("id").is(jobId)),
-                Update.update("paused", paused), JobDataMap.class);
-        JobDataMap data = mongoTemplate.findById(jobId, JobDataMap.class);
+        UpdateResult updateResult = jobDataMapService.updatePaused(jobId, paused);
+        JobDataMap data = jobDataMapService.findById(jobId);
         quartzService.pausedJob(data);
         return updateResult;
     }
